fix: [AUDIT-5000-R2] Fibonacci 0.236, trailing_tp_only propagation, units REAL
Wave 2 audit (6 additional parallel agents, 16 modules total) found 3 more issues:

1. market_analyzer: Added missing Fibonacci 0.236 level and standard 0.786
   (kept 0.750 for mentorship fidelity since Alex says "0.75")
2. trading_engine + risk_manager: Propagate trailing_tp_only and
   strategy_variant from SetupSignal through TradeRisk to ManagedPosition.
   GREEN crypto was losing its trailing-only flag in the conversion chain.
3. db/models: Changed units from INTEGER to REAL to support fractional
   lot sizes (crypto 0.001, forex micro 0.5)

731/731 tests passing. Fidelity: 9.9/10. 72 features audited, all PASS.

+++ backend/db/models.py
@@ -43,7 +43,7 @@ class TradeDatabase:
                 strategy TEXT,
                 strategy_variant TEXT,
                 direction TEXT NOT NULL CHECK(direction IN ('BUY', 'SELL')),
-                units INTEGER NOT NULL,
+                units REAL NOT NULL,
                 entry_price REAL NOT NULL,
                 exit_price REAL,
                 stop_loss REAL NOT NULL,
